Create the EmailDomainMatchingService
This service will be used to find an institution for a given email address. Matching will be performed using the email domain value objects. They are capable of matching an email address to their domain (with or without wildcards). The first match will yield the return of the associated institution.