Integrate the EmailDomainMatchingService
In doing so, an EmailAddressType and model where introduced. The model
was created over the existing domain layer EmailAddress VO because it
would be hard to verify the VO's correctness using validation
constraints. Read: we do not want validation annotations in the domain
layer.

The form is handled in the default controller. The controller also
verifies if the provided email address matches any of the configured
institutions. If it does not, a custom error message is produced. This
is a candidate for fixing up.