Remove UserRepository from GSSP
The UserRepository is removed to prevent storage handling in the
application. In order to be able to remove the repoository the
ADFS NameId (email address) is added to the GSSP NameId which
will be used in Stepup as SecondFactorIdentifier.

A timebased unique id is prepended to the id to have a unique id
per registration so multiple tokens with the same email address
could be used in Stepup.