Repair the behat registration test
The behat test is initiated from a non existant SP (configured as https://service_provider/saml/metadata)

That SP initiates the AuthNRequest. The gssp will handle the request using the mock azure MFA SP.

The test stops at the point where the GSSP sends back the SAML response to the initiating SP (which is not an actual SP). This SP can not be actually visited as it is non existent.