Repair the MockAzureMfaControllerTest
Using Panther required some changes to the test syntax

And because we now test in browser, the no-script blocks are no longer relied on. And a new identifier was required in the mock-acs.html.twig for each button. For now I've set the same value found on the noscript buttons on the normal ones.