Remove authentication template and update base
The base template no longer displays the auth and registration pages.
These pages should not be visited outside the context of an
authentication or registration SAML flow. Removing them makes the UI
less confusing.

The removal of the authentication template was warranted as no user
interaction is required at this point. The user email address is
provided in the SP authnrequest.