Improve error reporting for production
Bugfix: The GelfMessageToStringFormatter was not actually used, fixed
this by giving it a custom service name. This also required an
additional service definition, configuring the $formatter argument of
the service.

The prod monolog handlers have been update to more closely match the
stepup monolog configuration in use. Most notable is the addition of
the framework.php_errors.log config setting. This is added to the prod
monolog yaml in order to prevent logging op deprecation warnings.