Read original issuer from Scoping
Gateway will set scoping on the authnrequest, it contains the SP where
the original authnrequest originated from.

This value is used to judge if this is an RA authentication (vetting
procedure). If this is the case, the force authn is added to the
outgoing request.

This is deemed POC code, and should be reviewed carefully, any
additional defence in depth changes are welcome but might be postponed
until the team descides how to move forward with the POC.

https://www.pivotaltracker.com/story/show/171101458