Make email address client side validation work again.
By adding the 'type="email"' in the attr array, the <input> tag
effectively gets two type= attributes in the HTML, text and email,
this is invalid HTML but more importantly makes client side validation
not work.

The server side validation presents an ugly error so fixing this
client side is useful.