Fix crash when there is no institution found based on the user identity
The error now occurs due to the automtatic redirect based on the userattributes
in the SAML extention.
This will throw an error when there is no institution configured for the
users identity. It will yield a generic error page with the message that
the emailaddr of the user does not match any of the configured institutions

See: https://www.pivotaltracker.com/n/projects/1163646/stories/175121584