Replace node-sass with sass
Commit message from from my JS expert collegue Badlapje:

Node-sass is no longer the best practice to compile sass to css.  It has
some important drawbacks (such as requiring C and Python) and regularly
causes problems during the build.

In this change i've replaced it with the Sass package, an all-js
compilation package.  This has as an added benefit that it throws
deprecation warnings.