Hotfix the WebTests
The app_env override prevented the webtest from running in test env, as
the default is now always set to PROD. I fixed this by overriding the
parametesr.yaml for test.

Not the best solution, but works for now.

When @quartje's work is landing, and we start running the tests in a
dedicated test container, this work should be reverted partially.

Additionally I fixed an issue where the app-secret was set to the
app_env.