Allow metadata url keyrollover
This is needed in order to integrate an identity provider with the help
of a metadata url when doing a request. So an automatic key rollover
can take place and the IdP config is cached from the metadata url.

The update of the config is done in the callout itself and is not a
separate flow. This was done to prevent downtime because only one
certificate is in the metadata endpoint of the Microsoft identity
providers.

https://github.com/OpenConext/Stepup-AzureMFA/issues/194